backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
faiss-cpu
openai
pydantic
python-docx
PyPDF2
pandas
